+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changes
 -------
 * Add explicit bounds of [000, 366] to day of year component :code:`_parse_ordinal_date`, this adds the same limits to dates of the format YYYYDDD or YYYY-DDD when using :code:`parse_date`
 * Add explicit :code:`range_check` to :code:`BaseTimeBuilder`, call it explicitly before every build, this breaks compatibility with builders which have not been updated
+* Raise :code:`DayOutOfBoundsError` if calendary day exceeds number of days in calendar month
 
 Deprecation
 -----------